make WaypointNavigation the default one
