+++ field_friend/automations/navigation/__init__.py
@@ -1,8 +1,7 @@
 from .field_navigation import FieldNavigation, RowSegment
 from .implement_demo_navigation import ImplementDemoNavigation
-from .navigation import Navigation, WorkflowException
+from .navigation import Navigation, PathSegment, WorkflowException, WorkingSegment
 from .straight_line_navigation import StraightLineNavigation
-from .waypoint_navigation import PathSegment, WaypointNavigation, WorkingSegment
 
 __all__ = [
     'FieldNavigation',
@@ -11,7 +10,6 @@ __all__ = [
     'PathSegment',
     'RowSegment',
     'StraightLineNavigation',
-    'WaypointNavigation',
     'WorkflowException',
     'WorkingSegment',
 ]
