+++ field_friend/system.py
@@ -24,13 +24,7 @@ from .automations import (
     Puncher,
 )
 from .automations.implements import Implement, Recorder, Tornado, WeedingScrew
-from .automations.navigation import (
-    FieldNavigation,
-    ImplementDemoNavigation,
-    Navigation,
-    StraightLineNavigation,
-    WaypointNavigation,
-)
+from .automations.navigation import FieldNavigation, ImplementDemoNavigation, Navigation, StraightLineNavigation
 from .config import get_config
 from .hardware import Axis, FieldFriend, FieldFriendHardware, FieldFriendSimulation, TeltonikaRouter
 from .info import Info
@@ -252,11 +246,11 @@ class System(rosys.persistence.Persistable):
         self.field_navigation = FieldNavigation(self, first_implement).persistent() if self.gnss is not None else None
         self.implement_demo_navigation = ImplementDemoNavigation(self, first_implement).persistent() \
             if isinstance(self.field_friend.y_axis, Axis) else None
-        self.waypoint_navigation = WaypointNavigation(self, first_implement).persistent()
+        # self.waypoint_navigation = Navigation(self, first_implement).persistent()
         self.navigation_strategies = {n.name: n for n in [self.straight_line_navigation,
                                                           self.field_navigation,
                                                           self.implement_demo_navigation,
-                                                          self.waypoint_navigation,
+                                                          #   self.waypoint_navigation,
                                                           ] if n is not None}
         self.current_navigation = self.straight_line_navigation
