+++ tests/test_weeding.py
@@ -5,7 +5,7 @@ from rosys.testing import forward
 
 from field_friend import System
 from field_friend.automations.implements import Tornado, WeedingScrew
-from field_friend.automations.navigation import PathSegment, WaypointNavigation, WorkingSegment
+from field_friend.automations.navigation import PathSegment, StraightLineNavigation, WorkingSegment
 
 
 async def test_working_with_weeding_screw(system: System, detector: rosys.vision.DetectorSimulation):
@@ -95,6 +95,7 @@ async def test_weeding_screw_advances_when_there_are_no_plants(system: System, d
     assert detector.simulated_objects[1].category_name == 'weed'
     system.current_implement = system.implements['Weed Screw']
     system.current_navigation = system.straight_line_navigation
+    assert isinstance(system.current_navigation, StraightLineNavigation)
     system.current_navigation.length = 1.5
     assert isinstance(system.current_implement, WeedingScrew)
     system.current_implement.cultivated_crop = cultivated_crop
@@ -125,6 +126,7 @@ async def test_weeding_screw_advances_when_there_are_no_weeds_close_enough_to_th
     assert len(detector.simulated_objects) == 5
     system.current_implement = system.implements['Weed Screw']
     system.current_navigation = system.straight_line_navigation
+    assert isinstance(system.current_navigation, StraightLineNavigation)
     system.current_navigation.length = 1.5
     assert isinstance(system.current_implement, WeedingScrew)
     system.current_implement.cultivated_crop = 'maize'
@@ -151,7 +153,7 @@ async def test_implement_usage(system: System, detector: rosys.vision.DetectorSi
         ]
     system.current_navigation = system.waypoint_navigation
     system.current_implement = system.implements['Weed Screw']
-    assert isinstance(system.current_navigation, WaypointNavigation)
+    assert system.current_navigation is not None
     system.current_navigation.generate_path = generate_path  # type: ignore
     assert len(detector.simulated_objects) == 2
     system.automator.start()
@@ -177,6 +179,7 @@ async def test_work_offset_navigation(system: System, detector: rosys.vision.Det
 
 @pytest.mark.parametrize('system', ['u4'], indirect=True)
 async def test_tornado_removes_weeds_around_crop(system: System, detector: rosys.vision.DetectorSimulation):
+    assert isinstance(system.implements['Tornado'], Tornado)
     INNER_DIAMETER, OUTER_DIAMETER = system.field_friend.tornado_diameters(system.implements['Tornado'].tornado_angle)
     INNER_RADIUS = INNER_DIAMETER / 2
     OUTER_RADIUS = OUTER_DIAMETER / 2
@@ -187,8 +190,10 @@ async def test_tornado_removes_weeds_around_crop(system: System, detector: rosys
     detector.simulated_objects.append(rosys.vision.SimulatedObject(category_name='weed',
                                                                    position=rosys.geometry.Point3d(x=0.2, y=OUTER_RADIUS + 0.01, z=0)))
     targets = [
-        rosys.vision.SimulatedObject(category_name='weed', position=rosys.geometry.Point3d(x=0.2, y=INNER_RADIUS + 0.01, z=0)),
-        rosys.vision.SimulatedObject(category_name='weed',  position=rosys.geometry.Point3d(x=0.2, y=OUTER_RADIUS - 0.01, z=0))
+        rosys.vision.SimulatedObject(category_name='weed', position=rosys.geometry.Point3d(
+            x=0.2, y=INNER_RADIUS + 0.01, z=0)),
+        rosys.vision.SimulatedObject(
+            category_name='weed',  position=rosys.geometry.Point3d(x=0.2, y=OUTER_RADIUS - 0.01, z=0))
     ]
     detector.simulated_objects.extend(targets)
     system.current_implement = system.implements['Tornado']
@@ -213,12 +218,15 @@ async def test_tornado_drill_with_open_tornado(system: System, detector: rosys.v
     detector.simulated_objects.append(rosys.vision.SimulatedObject(category_name='weed',
                                                                    position=rosys.geometry.Point3d(x=0.4, y=MAX_RADIUS + 0.01, z=0)))
     targets = [
-        rosys.vision.SimulatedObject(category_name='weed', position=rosys.geometry.Point3d(x=0.4, y=MIN_RADIUS + 0.01, z=0)),
-        rosys.vision.SimulatedObject(category_name='weed',  position=rosys.geometry.Point3d(x=0.4, y=MAX_RADIUS - 0.01, z=0))
+        rosys.vision.SimulatedObject(
+            category_name='weed', position=rosys.geometry.Point3d(x=0.4, y=MIN_RADIUS + 0.01, z=0)),
+        rosys.vision.SimulatedObject(
+            category_name='weed',  position=rosys.geometry.Point3d(x=0.4, y=MAX_RADIUS - 0.01, z=0))
     ]
     detector.simulated_objects.extend(targets)
     system.current_implement = system.implements['Tornado']
     system.current_navigation = system.straight_line_navigation
+    assert isinstance(system.current_implement, Tornado)
     system.current_implement.tornado_angle = 180
     system.current_implement.drill_with_open_tornado = True
     system.automator.start()
@@ -239,6 +247,7 @@ async def test_tornado_skips_crop_if_no_weeds(system: System, detector: rosys.vi
                                                                    position=rosys.geometry.Point3d(x=0.6, y=0.0, z=0)))
     system.current_implement = system.implements['Tornado']
     system.current_navigation = system.straight_line_navigation
+    assert isinstance(system.current_implement, Tornado)
     system.current_implement.skip_if_no_weeds = True
     system.automator.start()
     await forward(until=lambda: system.automator.is_running)
